+++ packages/core/core-flows/src/rbac/steps/create-rbac-roles.ts
@@ -2,18 +2,34 @@ import { Modules } from "@medusajs/framework/utils"
 import { StepResponse, createStep } from "@medusajs/framework/workflows-sdk"
 import { IRbacModuleService } from "@medusajs/types"
 
+/**
+ * @ignore
+ * @featureFlag rbac
+ */
 export type CreateRbacRoleDTO = {
   name: string
   description?: string | null
   metadata?: Record<string, unknown> | null
 }
 
+/**
+ * @ignore
+ * @featureFlag rbac
+ */
 export type CreateRbacRolesStepInput = {
   roles: CreateRbacRoleDTO[]
 }
 
+/**
+ * @ignore
+ * @featureFlag rbac
+ */
 export const createRbacRolesStepId = "create-rbac-roles"
 
+/**
+ * @ignore
+ * @featureFlag rbac
+ */
 export const createRbacRolesStep = createStep(
   createRbacRolesStepId,
   async (data: CreateRbacRolesStepInput, { container }) => {