+++ www/apps/cloud/sidebar.mjs
@@ -29,6 +29,11 @@ export const sidebar = [
             path: "/comparison",
             title: "Cloud vs Self Hosting",
           },
+          {
+            type: "link",
+            title: "Command Palette",
+            path: "/command-palette",
+          },
         ],
       },
       {
@@ -51,6 +56,11 @@ export const sidebar = [
                 title: "Prerequisites",
                 path: "/projects/prerequisites",
               },
+              {
+                type: "link",
+                title: "Rename Repository",
+                path: "/projects/rename-repo-branch",
+              },
             ],
           },
           {
@@ -72,6 +82,11 @@ export const sidebar = [
                 type: "link",
                 title: "Environment Variables",
                 path: "/environments/environment-variables",
+              },
+              {
+                type: "link",
+                title: "Custom Domains",
+                path: "/environments/custom-domains",
               },
             ],
           },